feat: improve audit system architecture and eliminate code duplication
- Create shared audit utilities module to avoid code duplication
- Update audit logger to use shared utility functions
- Remove duplicated extractIpAddress, getEntityComplianceFlags methods
- Fix ID generation in data export to use UUID instead of Date.now()
- Add proper block scoping for switch cases
- Fix IP address extraction to handle missing request.ip property
- Improve audit middleware type safety
