+++ src/lib/audit/audit-middleware.ts
@@ -6,6 +6,7 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { getServerSession } from 'next-auth';
 import { getAuditLogger, AuditAction, AuditEntity, AuditEvent } from './audit-logger';
+import { extractIpAddress, getEntityComplianceFlags, inferActionFromMethod, inferClientType, extractAuditContext, createAuditMetadata } from './audit-utils';
 import { authOptions } from '@/lib/auth/auth-options';
 import { prisma } from '@/lib/prisma';
 
@@ -71,8 +72,8 @@ export function withAuditLogging<T extends any[]>(
     const baseEvent = {
       userId,
       organizationId,
-      action: options.action || inferActionFromMethod(method),
-      entity: options.entity || 'API',
+      action: options.action || (inferActionFromMethod(method) as AuditAction),
+      entity: options.entity || ('API' as AuditEntity),
       entityId: typeof entityId === 'string' ? entityId : undefined,
       resource: options.resource || path.split('/')[2] || 'unknown',
       method,
@@ -172,7 +173,7 @@ export function withDataAudit(
     handler: (req: NextRequest, ...args: T) => Promise<NextResponse>
   ) {
     return withAuditLogging(handler, {
-      action: action || inferActionFromMethod(req.method),
+      action: action || 'READ',
       entity,
       entityId: entityIdExtractor,
       resource: entity.toLowerCase(),
@@ -193,7 +194,7 @@ export function withDataAudit(
  */
 export function withPermissionAudit(
   requiredPermission: string,
-  entity: AuditEntity = 'API'
+  entity: AuditEntity = 'API_KEY'
 ) {
   return function<T extends any[]>(
     handler: (req: NextRequest, ...args: T) => Promise<NextResponse>
@@ -263,97 +264,15 @@ export function withComplianceAudit(
   };
 }
 
-// ============================================================================
-// UTILITY FUNCTIONS
-// ============================================================================
 
-function extractIpAddress(request: NextRequest): string {
-  return (
-    request.headers.get('x-forwarded-for')?.split(',')[0] ||
-    request.headers.get('x-real-ip') ||
-    request.ip ||
-    'unknown'
-  );
-}
 
-function inferActionFromMethod(method: string): AuditAction {
-  const methodMap: Record<string, AuditAction> = {
-    GET: 'READ',
-    POST: 'CREATE',
-    PUT: 'UPDATE',
-    PATCH: 'UPDATE',
-    DELETE: 'DELETE',
-  };
-  return methodMap[method.toUpperCase()] || 'READ';
-}
 
-function inferClientType(userAgent?: string): AuditEvent['clientType'] {
-  if (!userAgent) return 'unknown' as any;
-  
-  if (userAgent.includes('Mobile') || userAgent.includes('Android') || userAgent.includes('iPhone')) {
-    return 'mobile';
-  }
-  
-  if (userAgent.includes('curl') || userAgent.includes('Postman') || userAgent.includes('axios')) {
-    return 'api';
-  }
-  
-  return 'web';
-}
 
-function getEntityComplianceFlags(entity: AuditEntity): string[] {
-  const complianceMap: Record<string, string[]> = {
-    USER: ['SOX', 'GDPR', 'SOC2', 'HIPAA'],
-    RISK: ['SOX', 'SOC2', 'ISO27001'],
-    CONTROL: ['SOC2', 'ISO27001', 'NIST'],
-    COMPLIANCE_FRAMEWORK: ['SOX', 'SOC2', 'ISO27001'],
-    DOCUMENT: ['GDPR', 'SOC2', 'HIPAA'],
-    PAYMENT: ['PCI_DSS', 'SOX'],
-    SUBSCRIPTION: ['SOX', 'PCI_DSS'],
-    SYSTEM: ['SOC2', 'ISO27001'],
-    ORGANIZATION: ['SOX', 'GDPR', 'SOC2'],
-    API_KEY: ['SOC2', 'ISO27001'],
-    INTEGRATION: ['SOC2'],
-    ASSESSMENT: ['SOC2', 'ISO27001'],
-    REPORT: ['SOC2'],
-  };
-
-  return complianceMap[entity] || ['SOC2'];
-}
 
 // ============================================================================
 // AUDIT QUERY HELPERS
 // ============================================================================
 
-/**
- * Helper to extract common audit context from requests
- */
-export function extractAuditContext(req: NextRequest) {
-  return {
-    userAgent: req.headers.get('user-agent') || undefined,
-    ipAddress: extractIpAddress(req),
-    organizationId: req.headers.get('organization-id') || 'unknown',
-    sessionId: req.headers.get('x-session-id') || undefined,
-    requestId: req.headers.get('x-request-id') || `req_${Date.now()}`,
-    clientType: inferClientType(req.headers.get('user-agent') || undefined),
-    timestamp: new Date(),
-  };
-}
-
-/**
- * Helper to create audit metadata for specific operations
- */
-export function createAuditMetadata(
-  operation: string,
-  additionalData?: Record<string, any>
-): Record<string, any> {
-  return {
-    operation,
-    timestamp: new Date().toISOString(),
-    auditVersion: '1.0',
-    ...additionalData,
-  };
-}
 
 // ============================================================================
 // BATCH AUDIT OPERATIONS
@@ -416,8 +335,6 @@ export {
   withPermissionAudit,
   withSystemAudit,
   withComplianceAudit,
-  extractAuditContext,
-  createAuditMetadata,
   logBatchAuditEvents,
   logBulkOperation,
 };
